Move lib/api-keys.ts to device-api directory
