+++ src/compose/utils.ts
@@ -18,8 +18,7 @@ import {
 } from './types/service';
 
 import log from '../lib/supervisor-console';
-
-import * as apiKeys from '../lib/api-keys';
+import * as apiKeys from '../device-api/api-keys';
 
 export function camelCaseConfig(
 	literalConfig: ConfigMap,
